add : separate service calculation
