+++ api/Controllers/CommisionController.cs
@@ -1,4 +1,5 @@
 using Microsoft.AspNetCore.Mvc;
+using AvalphaTechnologies.CommissionCalculator.Dtos;
 
 namespace AvalphaTechnologies.CommissionCalculator.Controllers
 {
@@ -18,17 +19,6 @@ namespace AvalphaTechnologies.CommissionCalculator.Controllers
         }
     }
 
-    public class CommissionCalculationRequest
-    {
-        public int LocalSalesCount { get; set; }
-        public int ForeignSalesCount { get; set; }
-        public decimal AverageSaleAmount { get; set; }
-    }
 
-    public class CommissionCalculationResponse
-    {
-        public decimal AvalphaTechnologiesCommissionAmount { get; set; }
 
-        public decimal CompetitorCommissionAmount { get; set; }
-    }
 }
